Added a controller that does a refresh of the old token.

+++ controllers/auth/signout.js
@@ -1,7 +1,29 @@
 const { Token } = require("../../models");
+const jwt = require("jsonwebtoken");
+
+const { REFRESH_TOKEN_KEY } = process.env;
 
 const signout = async (req, res) => {
   const { refreshToken } = req.cookies;
+
+  if (!refreshToken) {
+    return res.code(401).json({
+      status: "error",
+      code: 401,
+      message: "Not authorized.",
+    });
+  }
+
+  const userData = jwt.verify(refreshToken, REFRESH_TOKEN_KEY);
+  const tokenFromDB = await Token.findOne({ owner: userData.id });
+
+  if (!userData || !tokenFromDB) {
+    return res.code(401).json({
+      status: "error",
+      code: 401,
+      message: "Not authorized.",
+    });
+  }
 
   const { _id } = await Token.findOne({ refreshToken });
   await Token.findByIdAndRemove(_id);
